ML/receipt-scanner/requirements.txt
fastapi
uvicorn
paddlepaddle==2.6.2
paddleocr==2.7.3
httpx
Pillow
numpy<2.0.0
pydantic
python-multipart
